fix(rules): resolve SC001 edge-case misfires
- Accept capitalized hyphenated first words with valid suffixes
- Exempt single-term all-caps headings like TL;DR from all-caps check
- Apply whole-token compound match on heading subsequent words
- Skip contextual all-caps terms used as lowercase bold prose
- Share isAcceptableHyphenatedCompound between heading and bold paths
